requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
pandas
scikit-learn
matplotlib
scipy
dotenv
tqdmcaptum
